MSXgl 1.2.12 - [Build tool] Added support to create program using Yamanooto mapper (8 MB) - [Build tool] Fixed NEO-8 mapper bank initialization - [ROM mapper] Created a header file to centralized all ROM mapper related features; Converted bank switching macros to inline functions (to have C type validation) - [System] Moved and renamed Peek/Poke functions (also converted from macros to inline functions) - [Print] Fixed parameters size for Print_DrawTextAtV() helper function - [Sample] Improved the Targets sample to support all the target medias - [Doc] Added source code documentation to several modules header

+++ engine/src/localize.h
@@ -48,4 +48,7 @@ inline void Loc_SetLanguage(u8 lang) { g_Loc_CurLang = (const c8**)(g_Loc_DataRo
 //
 // Parameters:
 //   text - Index of the text to retreive in the current language (must be between 0 and TEXT_MAX-1).
-inline const c8* Loc_GetText(u8 text) { return g_Loc_CurLang[text]; }
+//
+// Return:
+//   Pointer to the zero terminated string.
+inline const c8* Loc_GetText(u16 text) { return g_Loc_CurLang[text]; }
